fix jsx-load and sonic client of threegit
