+++ JumpscaleCore/clients/gedis/GedisClient.py
@@ -195,7 +195,7 @@ class GedisClient(JSConfigBase):
     @property
     def _threebot_me(self):
         if not self._threebot_me_:
-            self._threebot_me_ = j.tools.threebot.me.get(self.threebot_local_profile, needexist=False)
+            self._threebot_me_ = j.tools.threebot.me.get(self.threebot_local_profile, tname="me", needexist=False)
         return self._threebot_me_
 
     @property
