+++ JumpscaleCore/servers/threebot/ThreebotServer.py
@@ -67,6 +67,7 @@ class ThreeBotServer(j.baseclasses.object_config):
         self._openresty_server = None
         self._startup_cmd = None
         self._zdb = None
+        self._sonic = None
         self.threebot_server = None
         self.web = False
         self.ssl = False
@@ -186,7 +187,7 @@ class ThreeBotServer(j.baseclasses.object_config):
 
             gevent.sleep(day1)
 
-    def start(self, background=False, restart=False, packages=None):
+    def start(self, background=False, restart=False, packages=None, with_shell=True):
         """
 
         kosmos -p 'j.servers.threebot.default.start(background=True)'
@@ -197,7 +198,8 @@ class ThreeBotServer(j.baseclasses.object_config):
         :type packages: list of str
         ports & paths used for threebotserver
         see: {DIR_BASE}/code/github/threefoldtech/jumpscaleX_core/docs/3Bot/web_environment.md
-
+        :param with_shell: if set, will run j.shell() after a successful start, defaults to True
+        :type with_shell: bool
         """
         packages = packages or []
 
@@ -313,7 +315,8 @@ class ThreeBotServer(j.baseclasses.object_config):
 
             p = j.threebot.packages
 
-            j.shell()  # for now removed otherwise debug does not work
+            if with_shell:
+                j.shell()  # for now removed otherwise debug does not work
 
             forever = event.Event()
             try:
